Correct grasp offset, pick loop
It sometimes happened that the gripper collided with the table.
This was caused by the given position from the grasp.
We use an offset to solve this problem, so that the grasping position is
set a few centimeters in the negative direction of the approach vector.

The objects get picked in a loop. An object is grasped, the arm is extended
and the gripper opened. Afterwards the arm goes into home position and the loop
is repeated.

Also the bounding box, that defines which grasps are valid, was set closer
to the table in the z direction. This enables the robot to grasp smaller objects.