Use moveit pick pipeline without object
The moveit pick pipeline was recently changed to function without giving any object.
We use this, because we don't have any idea about the measuerements of the objects we grasp.

Also, the structure of the launch files was changed, so that launching pick_object_loop.launch brings up
gpd as well as the picking routine.