Clean up code
Removed unused functions and variables.
The spawnBoundingBox function is no longer needed, because we can grasp without
specifying any object.
Thus also the variables used to set the size of the bounding box are no longer needed.

Due to a recent change in gpd the frame_id is now set in the clustered_grasps message.
This is now used instead of setting the correct frame manually via parameter.